calculate min stops

+++ include/busCompany.h
@@ -29,7 +29,7 @@ public:
     double minDistance(const std::string& originStop, const std::string& destinyStop);
     std::list<std::pair<const Stop*, std::string>> minDistancePath(const std::string& originStop, const std::string& destinyStop);
 
-    double minStops(const std::string& originStop, const std::string& destinyStop);
+    int minStops(const std::string& originStop, const std::string& destinyStop);
     std::list<std::pair<const Stop*, std::string>> minStopsPath(const std::string& originStop, const std::string& destinyStop);
 };
